refined logic for user account view
